refactor: extract automations menu-item-content sub-components
